move the functions related to new account creation to newaccounts.cpp
